reorganizing tests into multiple files
