+++ test/setup.jl
@@ -1,0 +1,2 @@
+using RayTracer
+using Test
